Rename mui-themeable.js and mui-theme-provider.js to follow new naming conventions
